requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo;python_version<"3.9"
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
Flask==2.2.2
fuzzywuzzy==0.18.0
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Levenshtein==0.20.7
MarkupSafe==2.1.1
multiexit==1.5.0
packaging==21.3
protobuf==4.21.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.21.0
python-Levenshtein==0.20.7
pytube==12.1.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rapidfuzz==2.12.0
redis==4.3.4
requests==2.28.1
rsa==4.9
six==1.16.0
spotipy==2.20.0
sqlparse==0.4.2
typing-extensions==4.3.0
tzdata==2022.4
tzlocal==2.1
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
youtube-dl==2021.12.17
zipp==3.8.1
